Change range "allocated" for variables
Augmented rules are sorted by number after the dot first, this way rules
that end on zero (indicator of the end of rule) should appear in the
beggining of the set. Because variables were always negative numbers,
this could not be the case, resulting in wrong transitions/parsing
table. One example of this is grammar "S : ", which has transitions by
null character.